fix localStorage init
