+++ src/Settings.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	import { LocalStorage } from '@floaterest/sv';
+</script>
+
+<LocalStorage key="games" storage="{{}}" style="font-family: Consolas, monospace"/>
